chore: drizzle schema fixes

+++ apps/events/src/server/db/drizzle/schema/code.ts
@@ -1,6 +1,6 @@
 import { createId } from '@paralleldrive/cuid2';
 import { relations } from 'drizzle-orm';
-import { pgTable, varchar, double, int, unique } from 'drizzle-orm/pg-core';
+import { pgTable, varchar, numeric, integer, unique } from 'drizzle-orm/pg-core';
 import { tier } from './tier';
 import { ticket } from './ticket';
 
@@ -13,8 +13,8 @@ export const code = pgTable(
 		code: varchar('code', { length: 191 }).notNull(),
 		tierId: varchar('tierId', { length: 191 }).notNull(),
 		type: varchar('type', { length: 191 }).notNull(),
-		value: double('value').notNull(),
-		limit: int('limit').notNull(),
+		value: numeric('value').notNull(),
+		limit: integer('limit').notNull(),
 		notes: varchar('notes', { length: 191 }).notNull().default('')
 	},
 	(table) => {
