+++ apps/events/src/server/db/drizzle/schema/eventlocation.ts
@@ -1,14 +1,14 @@
 import { createId } from '@paralleldrive/cuid2';
 import { relations } from 'drizzle-orm';
-import { pgTable, varchar, double } from 'drizzle-orm/pg-core';
+import { pgTable, varchar, numeric } from 'drizzle-orm/pg-core';
 import { event } from './event';
 
 export const eventLocation = pgTable('EventLocation', {
 	id: varchar('id', { length: 128 })
 		.$defaultFn(() => createId())
 		.primaryKey(),
-	long: double('long').notNull(),
-	lat: double('lat').notNull(),
+	long: numeric('long').notNull(),
+	lat: numeric('lat').notNull(),
 	name: varchar('name', { length: 191 })
 });
